NP-28592 Ensure paging component is reusable
