+++ src/pages/search/registration_search/RegistrationSearch.tsx
@@ -1,13 +1,10 @@
 import { useTranslation } from 'react-i18next';
-import { Box, TablePagination, Typography } from '@mui/material';
-import { useHistory } from 'react-router-dom';
+import { Box, Typography } from '@mui/material';
 import { RegistrationSearchResults } from './RegistrationSearchResults';
 import { ListSkeleton } from '../../../components/ListSkeleton';
-import { ROWS_PER_PAGE_OPTIONS } from '../../../utils/constants';
-import { dataTestId } from '../../../utils/dataTestIds';
-import { SearchParam } from '../../../utils/searchHelpers';
-import { SearchResponse } from '../../../types/common.types';
 import { Registration } from '../../../types/registration.types';
+import { SearchListPaging } from '../SearchListPaging';
+import { SearchResponse } from '../../../types/common.types';
 
 interface RegistrationSearchProps {
   searchResults?: SearchResponse<Registration>;
@@ -16,20 +13,6 @@ interface RegistrationSearchProps {
 
 export const RegistrationSearch = ({ searchResults, isLoadingSearch }: RegistrationSearchProps) => {
   const { t } = useTranslation();
-  const history = useHistory();
-  const params = new URLSearchParams(history.location.search);
-
-  const resultsParam = params.get(SearchParam.Results);
-  const fromParam = params.get(SearchParam.From);
-
-  const rowsPerPage = (resultsParam && +resultsParam) || ROWS_PER_PAGE_OPTIONS[1];
-  const page = (fromParam && resultsParam && Math.floor(+fromParam / rowsPerPage)) || 0;
-
-  const updatePath = (from: string, results: string) => {
-    params.set(SearchParam.From, from);
-    params.set(SearchParam.Results, results);
-    history.push({ search: params.toString() });
-  };
 
   return (
     <Box gridArea="results">
@@ -38,16 +21,7 @@ export const RegistrationSearch = ({ searchResults, isLoadingSearch }: Registrat
       ) : searchResults && searchResults.hits.length > 0 ? (
         <>
           <RegistrationSearchResults searchResult={searchResults} />
-          <TablePagination
-            data-testid={dataTestId.startPage.searchPagination}
-            rowsPerPageOptions={ROWS_PER_PAGE_OPTIONS}
-            component="div"
-            count={searchResults.size}
-            rowsPerPage={rowsPerPage}
-            page={page}
-            onPageChange={(_, newPage) => updatePath((newPage * rowsPerPage).toString(), rowsPerPage.toString())}
-            onRowsPerPageChange={(event) => updatePath('0', event.target.value)}
-          />
+          <SearchListPaging totalCount={searchResults.size} />
         </>
       ) : (
         <Typography>{t('common.no_hits')}</Typography>
